+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * `c_across()` and `across()` key deparsing not confused by long calls (#5883).
 
+* `across()` handles named selections (#5207).
+
 # dplyr 1.0.6
 
 * `add_count()` is now generic (#5837).